HH-218702 move NamedDataSource to nab-jdbc-common module
