+++ nab-telemetry-jdbc/src/main/java/ru/hh/nab/telemetry/jdbc/internal/NabTelemetryDataSource.java
@@ -9,9 +9,9 @@ import io.opentelemetry.instrumentation.jdbc.internal.ThrowingSupplier;
 import java.sql.Connection;
 import java.sql.SQLException;
 import javax.sql.DataSource;
-import ru.hh.nab.datasource.NamedDataSource;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.datasource.DelegatingDataSource;
+import ru.hh.nab.jdbc.common.datasource.NamedDataSource;
 import ru.hh.nab.telemetry.jdbc.internal.model.NabDataSourceInfo;
 import ru.hh.nab.telemetry.jdbc.internal.model.NabDbInfo;
 import ru.hh.nab.telemetry.jdbc.internal.model.NabDbRequest;
